+++ src/engine/execcmd.h
@@ -110,6 +110,6 @@ int is_raw_command_request( LIST * shell );
  * are under the specified length limit.
  */
 int check_cmd_for_too_long_lines( char const * command, size_t max,
-    int * const error_length, int * const error_max_length );
+    size_t * const error_length, size_t * const error_max_length );
 
 #endif
